FEXCore: Support forcing all allocations to the upper 48-bits
This is a debug feature that is not entirely stable but is super useful
when trying to profile FEX's memory allocations on 64-bit applications.

Typically FEX just mixes its own allocations with the guest
application's allocations for 64-bit, but we can pass the 48-bit
allocation to our memory allocator and force everything in to that space
with very little effort. This sets the jemalloc hooks when enabled which
adds a little bit of overhead.

Once the option is enabled you can look at pmap -x <pid> to see the
separation between FEX and guest allocations since all FEX allocations
end up in the high bits after VDSO is loaded.

Very handy!