Linux: Converts passthrough syscalls to generator script
After adding the past few Linux versions of syscalls I noticed that this
could be better.

Walks through all of our syscalls and converts them to a json format
that is parsed and then generated at compile time.

This also fixes some mislabeling of system calls that were getting
inlined on accident.

This really helps clean up the syscall implementation, all syscalls that
can get passed through to the host syscall handler is no longer visible
in the implementation.